Create mixin base class
