+++ safe/impact_reports/building_exposure_report_mixin.py
@@ -1,32 +1,22 @@
-__author__ = 'christian@kartoza.com'
+# coding=utf-8
+"""
+InaSAFE Disaster risk assessment tool developed by AusAid -
+**Impact Function Manager**
+
+Contact : ole.moller.nielsen@gmail.com
+
+.. note:: This program is free software; you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation; either version 2 of the License, or
+     (at your option) any later version.
+"""
 
 from safe.utilities.i18n import tr
 from safe.common.utilities import format_int
-from safe.common.tables import Table, TableRow
+from safe.impact_reports.report_mixin_base import ReportMixin
 
 
-class BuildingExposureReportMixin(object):
-
-    def generate_html_report(self, question, affected_buildings, buildings):
-        """Breakdown by building type.
-
-        :param question: The impact question.
-        :type question: basestring
-
-        :param affected_buildings: The affected buildings
-        :type affected_buildings: OrderedDict
-
-        :param buildings: The buildings and totals.
-        :type buildings: dict
-
-        :returns: The report in .
-        :rtype: list
-        """
-        return self.parse_to_html(
-            self.generate_report(
-                question,
-                affected_buildings,
-                buildings))
+class BuildingExposureReportMixin(ReportMixin):
 
     def generate_report(self, question, affected_buildings, buildings):
         """Breakdown by building type.
@@ -246,47 +236,3 @@ class BuildingExposureReportMixin(object):
                 if current_usage.lower() == usage.lower():
                     count += category_breakdown[current_usage]
         return count
-
-    def notes(self):
-        """Additional notes to be used.
-
-        :return: The notes to be added to this report
-
-        ..Notes:
-        Notes are very much specific to IFs so it is expected that this method
-        is overwritten in the IF if needed.
-        """
-        return []
-
-    def parse_to_html(self, report):
-        """Convert a json compatible list of results to a tabulated version.
-
-        :param report: A json compatible reoprt
-        :type report: list
-
-        :returns: Returns a tabulated version of the report
-        :rtype: basestring
-        """
-        tabulated_report = []
-        for row in report:
-            row_template = {
-                'content': '',
-                'condition': True,
-                'arguments': (),
-                'header': False
-            }
-            row_template.update(row)
-            if not row_template['condition']:
-                continue
-            if row_template['arguments']:
-                content = row_template['content'] % row_template['arguments']
-            else:
-                content = row_template['content']
-            if row_template['header']:
-                table_row = TableRow(content, header=True)
-            else:
-                table_row = TableRow(content)
-            tabulated_report.append(table_row)
-
-        html_tabulated_report = Table(tabulated_report).toNewlineFreeString()
-        return html_tabulated_report
